refactor(extensions): decompose near module + harden account-id validation
Addresses Firat's review on #4661:
- Decompose the 1216-line near.rs into near/{mod,rpc,input,tests}.rs so each
  file is well under the 1000-line threshold (mod 368, rpc 224, input 123,
  tests 651). Public API is unchanged.
- Add NEAR account-id charset validation (is_valid_near_account_id): 2-64 chars,
  lowercase alphanumeric with non-leading/trailing/repeating '.', '-', '_'
  separators. Length-only checks admitted IDs that failed opaquely at the RPC
  layer.
- Fill test gaps: account-id charset accept/reject, ft_balances partial-failure
  (try_join_all short-circuit -> OperationFailed), and intents_quote egress
  error -> NetworkDenied.
- Flag EXPERIMENTAL_tx_status instability with a tracked TODO on the method.
